move context and utils to src
