+++ client/src/utils/fetchWithAuth.ts
